move back initFile class member from SWTEncoding Class
